fix up button in steps
